move-only: Add util/hash_type
Can be reviewed with --color-moved=dimmed-zebra
